git-svn-id: svn://10.153.116.2/PAM@1583 cb18dcf6-275f-e045-ae08-2e8c34a2df07

+++ functions/PAM/LoadTcspc.m
@@ -424,7 +424,7 @@ switch (Type)
         if ~isempty(MI_Bins) && MI_Bins>1 %%% Sets number of MI bins to value from .set file (or fixed)
             FileInfo.MI_Bins = MI_Bins;
         else %%% Reads highest used MI and usen 2^n bins
-            usedMI = max(cellfun(@max,TcspcData.MI(~cellfun(@isempty,TcspcData.MI))));
+            usedMI = double(max(cellfun(@max,TcspcData.MI(~cellfun(@isempty,TcspcData.MI)))));
             % usedMI = max(cellfun(@numel,cellfun(@unique,TcspcData.MI(~cellfun(@isempty,TcspcData.MI)),'UniformOutput',false)));  
             FileInfo.MI_Bins = 2^(ceil(log2(usedMI)));
         end
